requirements.txt
Django
django-extensions
djangorestframework
djangorestframework-simplejwt
drf-spectacular